package structure and book details screen flow handling and UIState
