+++ app/src/main/java/com/caminaapps/bookworm/app/BookwormApplication.kt
@@ -1,0 +1,17 @@
+package com.caminaapps.bookworm.app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class BookwormApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        initLogger()
+    }
+
+    private fun initLogger() {
+        Timber.plant(Timber.DebugTree())
+    }
+}
