--- app/src/main/java/com/caminaapps/bookworm/core/data/local/AppDatabase.kt
@@ -1,28 +1,0 @@
-package com.caminaapps.bookworm.core.data.local
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-
-
-@Database(
-    entities = [BookEntity::class],
-    version = 1,
-    exportSchema = false
-)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun bookDao(): BookDao
-
-    companion object {
-        private const val databaseName = "bookworm-db"
-
-        fun getInstance(context: Context): AppDatabase {
-            return Room.databaseBuilder(
-                context,
-                AppDatabase::class.java,
-                databaseName
-            ).build()
-        }
-    }
-}
