+++ app/src/main/java/com/caminaapps/bookworm/core/data/network/interceptor/LoggingInterceptor.kt
@@ -1,0 +1,11 @@
+package com.caminaapps.bookworm.core.data.network.interceptor
+
+import okhttp3.logging.HttpLoggingInterceptor
+import timber.log.Timber
+import javax.inject.Inject
+
+class LoggingInterceptor @Inject constructor() : HttpLoggingInterceptor.Logger {
+    override fun log(message: String) {
+        Timber.d(message)
+    }
+}
