--- app/src/main/java/com/caminaapps/bookworm/core/data/remote/interceptor/LoggingInterceptor.kt
@@ -1,11 +1,0 @@
-package com.caminaapps.bookworm.core.data.remote.interceptor
-
-import okhttp3.logging.HttpLoggingInterceptor
-import timber.log.Timber
-import javax.inject.Inject
-
-class LoggingInterceptor @Inject constructor() : HttpLoggingInterceptor.Logger {
-    override fun log(message: String) {
-        Timber.d(message)
-    }
-}
