+++ app/src/main/java/com/caminaapps/bookworm/core/domain/repository/BookRepository.kt
@@ -6,8 +6,8 @@ import kotlinx.coroutines.flow.Flow
 
 interface BookRepository {
 
-    fun getAllBooks(): Flow<List<Book>>
-    fun getBookDetails(id: String): Flow<Book?>
+    fun getAllBooksStream(): Flow<List<Book>>
+    fun getBookDetailsStream(id: String): Flow<Book?>
 
     suspend fun saveBook(book: Book)
     suspend fun deleteBook(id: String)
