--- app/src/main/java/com/caminaapps/bookworm/core/presentation/BookwormApplication.kt
@@ -1,17 +1,0 @@
-package com.caminaapps.bookworm.core.presentation
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class BookwormApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        initLogger()
-    }
-
-    private fun initLogger() {
-        Timber.plant(Timber.DebugTree())
-    }
-}
