--- app/src/main/java/com/caminaapps/bookworm/core/presentation/main/MainActivity.kt
@@ -1,25 +1,0 @@
-package com.caminaapps.bookworm.core.presentation.main
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
-import androidx.compose.material.MaterialTheme
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.Modifier
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import dagger.hilt.android.AndroidEntryPoint
-
-@ExperimentalComposeUiApi
-@ExperimentalPermissionsApi
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            setContent {
-                MainScreen(Modifier.background(MaterialTheme.colors.background))
-            }
-        }
-    }
-}
