--- app/src/main/java/com/caminaapps/bookworm/core/presentation/screens/settings/SettingsScreen.kt
@@ -1,12 +1,0 @@
-package com.caminaapps.bookworm.core.presentation.screens.settings
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.hilt.navigation.compose.hiltViewModel
-
-@Composable
-fun SettingsScreen(
-    viewModel: SettingsViewModel = hiltViewModel()
-) {
-    Text("- Settings -")
-}
