--- app/src/main/java/com/caminaapps/bookworm/core/presentation/screens/settings/SettingsViewModel.kt
@@ -1,10 +1,0 @@
-package com.caminaapps.bookworm.core.presentation.screens.settings
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class SettingsViewModel @Inject constructor(
-
-) : ViewModel() {}
