--- app/src/main/java/com/caminaapps/bookworm/core/presentation/screens/wishlist/WishlistScreen.kt
@@ -1,12 +1,0 @@
-package com.caminaapps.bookworm.core.presentation.screens.wishlist
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.hilt.navigation.compose.hiltViewModel
-
-@Composable
-fun WishlistScreen(
-    viewModel: WishlistViewModel = hiltViewModel()
-) {
-    Text("- Wishlist -")
-}
