--- app/src/main/java/com/caminaapps/bookworm/core/presentation/screens/wishlist/WishlistViewModel.kt
@@ -1,10 +1,0 @@
-package com.caminaapps.bookworm.core.presentation.screens.wishlist
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class WishlistViewModel @Inject constructor(
-
-) : ViewModel() {}
