+++ app/src/main/java/com/caminaapps/bookworm/di/LocalDataSourceModule.kt
@@ -1,8 +1,8 @@
 package com.caminaapps.bookworm.di
 
 import android.content.Context
-import com.caminaapps.bookworm.core.data.local.AppDatabase
-import com.caminaapps.bookworm.core.data.local.BookDao
+import com.caminaapps.bookworm.core.data.database.AppDatabase
+import com.caminaapps.bookworm.core.data.database.BookDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
