+++ app/src/main/java/com/caminaapps/bookworm/di/RemoteDataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.caminaapps.bookworm.di
 
-import com.caminaapps.bookworm.core.data.remote.OpenLibraryAPI
-import com.caminaapps.bookworm.core.data.remote.interceptor.LoggingInterceptor
+import com.caminaapps.bookworm.core.data.network.OpenLibraryAPI
+import com.caminaapps.bookworm.core.data.network.interceptor.LoggingInterceptor
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import dagger.Module
 import dagger.Provides
