+++ app/src/main/java/com/caminaapps/bookworm/features/book/add/AddBookViewModel.kt
@@ -1,0 +1,11 @@
+package com.caminaapps.bookworm.features.book.add
+
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class AddBookViewModel @Inject constructor(
+    savedStateHandle: SavedStateHandle,
+) : ViewModel() {}
