+++ app/src/main/java/com/caminaapps/bookworm/features/settings/SettingsScreen.kt
@@ -1,0 +1,12 @@
+package com.caminaapps.bookworm.features.settings
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
+
+@Composable
+fun SettingsScreen(
+    viewModel: SettingsViewModel = hiltViewModel()
+) {
+    Text("- Settings -")
+}
