+++ app/src/main/java/com/caminaapps/bookworm/features/settings/SettingsViewModel.kt
@@ -1,0 +1,10 @@
+package com.caminaapps.bookworm.features.settings
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class SettingsViewModel @Inject constructor(
+
+) : ViewModel() {}
