+++ app/src/main/java/com/caminaapps/bookworm/features/wishlist/WishlistViewModel.kt
@@ -1,0 +1,10 @@
+package com.caminaapps.bookworm.features.wishlist
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class WishlistViewModel @Inject constructor(
+
+) : ViewModel() {}
